Added loading screen

+++ lambda/yelpSearch/index.js
@@ -12,8 +12,8 @@ exports.handler = (e, context, callback) => {
 	var yelpSearch = new Promise((resolve, reject) => {
 		var data = yelp.search({
 			term: decodeURI(e.keyword), 
-			longitude: parseInt(e.longitude), 
-			latitude: parseInt(e.latitude),
+			longitude: parseFloat(e.longitude), 
+			latitude: parseFloat(e.latitude),
 			limit: 25,
 			radius: 40000
 		})
